Type promotion in haskell
Like in Haskell, the arithmetic operations of C++ always act on
identical types (ex: 2 + 2 or 2.0 + 2.0, but not 2 + 2.0).

Unlike Haskell, C++ automatically promotes the arguments of arithmetic
operations to the most general type. Lest it be thought that C++ is
superior to Haskell in this regard, here is an implementation of type
promotion in Haskell.